bundle:: add a common module for shared code
